readme and code clean of clustering
